bot_tel/requirements.txt

python-telegram-bot==20.3
pytz
python-dotenv
